refactoring, adding/moving low-level functions to internal modules
