+++ Graphics/GL/Low/Internal.hs
@@ -1,0 +1,13 @@
+module Graphics.GL.Low.Internal (
+    module Graphics.GL.Low.Internal.Types,
+    module Graphics.GL.Low.Internal.GLEnum,
+    module Graphics.GL.Low.Internal.BufferObject,
+    module Graphics.GL.Low.Internal.Shader
+    ) where
+
+import Graphics.GL.Low.Internal.Types
+import Graphics.GL.Low.Internal.GLEnum
+import Graphics.GL.Low.Internal.BufferObject
+import Graphics.GL.Low.Internal.Shader
+
+
