+++ lowgl.cabal
@@ -65,8 +65,12 @@ library
                        Graphics.GL.Low.Texture
                        Graphics.GL.Low.VAO
                        Graphics.GL.Low.VertexAttrib
-                       Graphics.GL.Low.Common
+                       Graphics.GL.Low.Internal
                        Graphics.GL.Low.Internal.Types
+                       Graphics.GL.Low.Internal.GLEnum
+                       Graphics.GL.Low.Internal.Common
+                       Graphics.GL.Low.Internal.BufferObject
+                       Graphics.GL.Low.Internal.Shader
   
   -- Modules included in this library but not exported.
   -- other-modules:       
@@ -79,8 +83,7 @@ library
                        transformers >= 0.4 && < 0.5,
                        vector >=0.10 && <0.11, 
                        linear >= 1.16 && <1.17, 
-                       gl >=0.5 && <0.8, 
-                       data-default
+                       gl >=0.5 && <0.8
   
   -- Directories containing source files.
   -- hs-source-dirs:
